egi/requirements.txt
argon2-cffi==19.2.0
asn1crypto==1.2.0
boto3==1.10.19
botocore==1.13.19
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Django==2.2.7
django-appconf==1.0.3
django-axes==5.0.19
django-health-check==3.11.1
django-ipware==2.1.0
django-ratelimit==2.0.0
django-storages==1.7.2
django-webpack-loader==0.6.0
djangorestframework==3.10.3
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
mysqlclient==1.4.5
numpy==1.17.4
pandas==0.25.3
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
retrying==1.3.3
s3transfer==0.2.1
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
watchtower==0.7.3
Werkzeug==0.16.0
whitenoise==4.1.4
xlrd==1.2.0
